Add support for multiple departments per vote
This is needed to exclude departments which are subsumed in
other departments' votes like Civilian Secretariat in Police
when aggregating the ene pivot data across votes.

The idea here is that only one department can be the primary
department for a vote. Zero or more can be non-primary
departments. To find the name of the vote to find vote-related
data in ckan, for example, you find the primary department
with the same vote number in the same government.