Serialise decimals as floats
This using this data at the moment expect floats, not decimal strings.
They humanise the numbers before presentation so weirdness will get rounded away.

It might be good to serialise decimals as strings if we provide this in a
pubic API at some point though, then we can review the best course of action